requirements.txt
opencv-python
matplotlib
tqdm
mediapipe
protobuf
pyaml
easydict
tensorboardX
timm==0.5.4
einops
mmcv
imageio
open3d
terminaltables
simpleicp
pymatting
gdown
onnxruntime
seaborn